test(detailing): stop re-running the flagship once per coverage assertion
The coverage-invariant file called runDetailing five times on the same
inputs — once per test — and each call is 8,6 s on the 408-member
fixture. Measured, that was 43 s of a 47 s file, and it is what pushed
the CI worker past the reporter's RPC budget while every assertion
passed.

The run moves into a helper that memoises it per module. Vitest isolates
modules per test file, so the memo never crosses a file boundary and no
test can observe another file's state; within a file the result is
read-only by contract, which is what the previous file already relied on
for its solve-and-design step. The helper computes nothing the tests are
about — it calls the production entry points with production arguments
and returns what they gave back.

The assertions then split by concern rather than sitting in one task.
Outcome population needs only solve and design, so it no longer pays for
detailing at all. Ownership and output are separate questions and each
pays for one detailing run.

Same fixture, same production path, same seven tests, same twelve
assertions, none weakened. Longest worker task 47,2 s to 11,9 s; total
work across the group 43 s to 25,8 s; wall time 47,2 s to 13,8 s, over
three runs each.
